pyproject.toml
[project]
name = "zmk"
description = "A program to help set up ZMK Firmware"
readme = "README.md"
license = { file = "LICENSE" }
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Topic :: Software Development",
    "Topic :: Utilities",
]
requires-python = ">=3.10"
dependencies = [
    "platformdirs >= 3.0.0",
    "rich >= 13.6.0",
    "ruamel.yaml >= 0.18.6",
    "shellingham >= 1.5.3",
    "typer >= 0.12.0",
    "typing_extensions >= 3.7.4.3",
    "west >= 1.2.0",
]
dynamic = ["version"]

[project.urls]
Documentation = "https://zmk.dev/docs"
"Source Code" = "https://github.com/zmkfirmware/zmk_cli/"
"Issue Tracker" = "https://github.com/zmkfirmware/zmk_cli/issues/"
Chat = "https://zmk.dev/community/discord/invite"

[project.scripts]
zmk = "zmk.main:app"

[build-system]
requires = ["setuptools", "setuptools-scm"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["zmk"]

[tool.setuptools_scm]
write_to = "zmk/_version.py"
